1.1.2.1
 1.1.2.1 - Set SCE to AUX
 * [BUG 1.1.2.0a] TweakScale
 * [NEW][BUG 1.1.2.1a] AmpYear doesn't seem to recognize ODFC
 * BUG 1.1.2.0b] Kerbalism is not compatible with ODFC - Kerbalism developers have chosen to not integrate since they have their own version.
 * [D][BUG 1.1.2.1b] Found the problem, Bill left a torque converter, which was set to moar, in the innards. Plus Val thought it would be fun to hide her signature in the code of the FCOS (Fuel Cell Operating System). In doing so, she overwrote the code to remember the editor settings when translating over to flight. This has been fixed by moving her signature into another mod's code. She's still happy.
 * Updated to automated build (deploy.bat,buildRelease.bat/Assembly.tt)
