--- GameData/OnDemandFuelCells/MM-Patches/Copy/MiningExpansion.cfg
@@ -1,177 +1,0 @@
-// MiningExpansion.cfg v1.2
-// Stock Mining Enhancement (MiningExpansion)
-// On Demand Fuel Cells Refueled
-// created: 27 Aug 19
-// updated: 27 Aug 19
-
-// NOTE:wip
-
-+PART[SMX_Size0Generator]:NEEDS[MiningExpansion,ODFC]:FOR[ODFC]
-{
-	@name ^= :^:ODFC:
-	@title ^= :$: <color=green> ODFC</color>:
-	@description ^= :$: <color=green> ODFC version. </color>:
-	@tags ^= :$:  odfc demand:
-
-	!MODULE[ModuleResourceConverter] {}
-	MODULE {
-		name = ODFC
-		MODE
-		{
-			MaxEC = 16
-			FUELS
-			{
-				MonoPropellant = 0.016
-			}
-		}
-		MODE:NEEDS[CommunityResourcePack]
-		{
-			MaxEC = 16
-			FUELS
-			{
-				Hydrogen = 0.064736	// 0.272984334
-				Oxygen = 0.032368		// 0.137359451
-			}
-			BYPRODUCTS
-			{
-				Water = 0.0000520721	// 0.000220822
-			}
-		}
-		MODE // from part (ratio correct)
-		{
-			MaxEC = 6
-			FUELS {
-				LiquidFuel = 0.00675
-				Oxidizer = 0.00825
-			}
-		}
-		MODE
-		{
-			MaxEC = 16
-			FUELS {
-				LiquidFuel = 0.0001275
-				IntakeAir = 0.0204
-			}
-		}
-	}
-}
-
-+PART[SMX_Size1Generator]:NEEDS[MiningExpansion,ODFC]:FOR[ODFC]
-{
-	@name ^= :^:ODFC:
-	@title ^= :$: <color=green> ODFC</color>:
-	@description ^= :$: <color=green> ODFC version. </color>:
-	@tags ^= :$:  odfc demand:
-
-	!MODULE[ModuleResourceConverter] {}
-	MODULE {
-		name = ODFC
-		MODE
-		{
-			MaxEC = 16
-			FUELS
-			{
-				MonoPropellant = 0.016
-			}
-		}
-		MODE:NEEDS[CommunityResourcePack]
-		{
-			MaxEC = 16
-			FUELS
-			{
-				Hydrogen = 0.064736	// 0.272984334
-				Oxygen = 0.032368		// 0.137359451
-			}
-			BYPRODUCTS
-			{
-				Water = 0.0000520721	// 0.000220822
-			}
-		}
-		MODE // from part (ratio correct)
-		{
-			MaxEC = 50
-			FUELS {
-				LiquidFuel = 0.05625
-				Oxidizer = 0.06875
-			}
-		}
-		MODE
-		{
-			MaxEC = 16
-			FUELS {
-				LiquidFuel = 0.0001275
-				IntakeAir = 0.0204
-			}
-		}
-	}
-}
-
-+PART[SMX_Mk2Generator]:NEEDS[MiningExpansion,ODFC]:FOR[ODFC]
-{
-	@name ^= :^:ODFC:
-	@title ^= :$: <color=green> ODFC</color>:
-	@description ^= :$: <color=green> ODFC version. </color>:
-	@tags ^= :$:  odfc demand:
-
-	!MODULE[ModuleResourceConverter],0 {}
-	MODULE,0
- 	{
-		name = ODFC
-		MODE
-		{
-			MaxEC = 16
-			FUELS
-			{
-				MonoPropellant = 0.016
-			}
-		}
-		MODE:NEEDS[CommunityResourcePack]
-		{
-			MaxEC = 16
-			FUELS
-			{
-				Hydrogen = 0.064736
-				Oxygen = 0.032368
-			}
-			BYPRODUCTS
-			{
-				Water = 0.0000520721
-			}
-		}
-		MODE // from part (ratio correct)
-		{
-			MaxEC = 75
-			FUELS {
-				LiquidFuel = 0.084375
-				Oxidizer = 0.103125
-			}
-		}
-		MODE
-		{
-			MaxEC = 16
-			FUELS {
-				LiquidFuel = 0.0001275
-				IntakeAir = 0.0204
-			}
-		}
-	}
-
-	!MODULE[ModuleResourceConverter],1 {}
-	MODULE:NEEDS[MiningExpansion],1 {
-		name = ODFC
-		MODE//:NEEDS[MiningExpansion]
-		{
-			MaxEC = 75
-			FUELS
-			{
-				Ore = 0.1875
-			}
-			BYPRODUCTS
-			{
-				//LiquidFuel = 0
-			}
-		}
-}
-
-// CC BY-NC-SA-4.0
-// zer0Kerbal 4x4cheesecake linuxgurugamer
