--- GameData/OnDemandFuelCells/MM-Patches/Copy/SolidFuelCell.cfg
@@ -1,31 +1,0 @@
-// SolidFuelCell.cfg v1.1
-// On Demand Fuel Cells Refueled
-// created: 29 Aug 19
-// updated: 03 Sep 19
-
-+PART[solid_fuelcell]:NEEDS[ODFC,SolidFuelCell]:BEFORE[ODFC]
-{
-	@name ^= :^:ODFC:
-	@cost += 50
-	@title ^= :$: <#ADFF2F> ODFC</color>:
-	@description ^= :$: <#ADFF2F><i> ODFC version.</i> </color>:
-	@mass += 0.01
-	@tags ^= :$:  odfc demand:
-
-	!MODULE[ModuleResourceConverter] {}
-	MODULE
-	{
-		name = ODFC
-		MODE
-		{
-			MaxEC = 1.5
-			FUELS
-			{
-				SolidFuel = 0.00375
-			}
-		}
-	}
-}
-
-// CC BY-NC-SA-4.0
-// zer0Kerbal
