--- GameData/OnDemandFuelCells/MM-Patches/Copy/StockFuelCells.cfg
@@ -1,116 +1,0 @@
-// Stock Fuel Cells v1.7
-// On Demand Fuel Cells Refueled
-// created: 13 Aug 19
-// updated: 03 Sep 19
-
-+PART[FuelCell]:NEEDS[ODFC]:BEFORE[ODFC]
-{
-	@name ^= :^:ODFC:
-	@cost += 50
-	@title ^= :$: <#ADFF2F> ODFC</color>:
-	@description ^= :$: <#ADFF2F><i> ODFC version.</i> </color>:
-	@mass += 0.01
-	@tags ^= :$:  odfc demand:
-
-	!MODULE[ModuleResourceConverter] {}
-	MODULE,0
-	{
-		name = ODFC
-		MODE
-		{
-			MaxEC = 5
-			FUELS
-			{
-				MonoPropellant = 0.005
-			}
-		}
-		MODE
-		{
-			MaxEC = 5
-			FUELS
-			{
-				LiquidFuel = 0.005625
-				Oxidizer = 0.0152777777777778
-			}
-		}
-		MODE:NEEDS[CommunityResourcePack]
-		{
-			MaxEC = 5
-			FUELS
-			{
-				Hydrogen = 1.04264849617275
-				Oxygen = 0.788621982550263
-			}
-			BYPRODUCTS
-			{
-				Water = 0.000843419221524059
-			}
-		}
-		MODE
-		{
-			MaxEC = 5
-			FUELS
-			{
-				LiquidFuel = 0.000425
-				IntakeAir = 0.102
-			}
-		}
-	}
-}
-
-+PART[FuelCellArray]:NEEDS[ODFC]:BEFORE[ODFC]
-{
-	@name ^= :^:ODFC:
-	@cost += 50
-	@title ^= :$: <#ADFF2F> ODFC</color>:
-	@description ^= :$: <#ADFF2F><i> ODFC version.</i> </color>:
-	@mass += 0.01
-	@tags ^= :$:  odfc demand:
-
-	!MODULE[ModuleResourceConverter] {}
-	MODULE {
-		name = ODFC
-		MODE
-		{
-			MaxEC = 32
-			FUELS
-			{
-				MonoPropellant = 0.032
-			}
-		}
-		MODE
-		{
-			MaxEC = 32
-			FUELS
-			{
-				LiquidFuel = 0.036
-				Oxidizer = 0.0977777777777778
-			}
-		}
-		MODE:NEEDS[CommunityResourcePack]
-		{
-			MaxEC = 32
-			FUELS
-			{
-				Hydrogen = 6.67295037550558
-				Oxygen = 5.04718068832168
-			}
-			BYPRODUCTS
-			{
-				Water = 0.00539788301775398
-			}
-		}
-		MODE
-		{
-			MaxEC = 32
-			FUELS
-			{
-				LiquidFuel = 0.00272
-				IntakeAir = 0.6528
-			}
-		}
-	}
-}
-
-// CC BY-NC-SA-4.0
-// zer0Kerbal
