--- GameData/OnDemandFuelCells/MM-Patches/FuturepatchesWIP.txt
@@ -1,113 +1,0 @@
-// Stock Fuel Cells v1.7
-// On Demand Fuel Cells Refueled
-// created: 13 Aug 19
-// updated: 02 Sep 19
-
-+PART[crewCabin,mk1-3pod,landerCabinSmall,mk1pod,mk1pod_v2,mk2LanderCabin,mk2LanderCabin_V2,mk3Cockpit_Shuttle,cupola,mk2Pod]:NEEDS[ODFC]:BEFORE[ODFC]
-{
-	@name ^= :^:ODFC:
-	@cost += 50
-	@title ^= :$: <color=green> ODFC</color>:
-	@description ^= :$: <color=green><i> ODFC version.</i> </color>:
-	@mass += 0.010
-	@tags ^= :$:  odfc demand:
-
-	!MODULE[ODFC] {}
-	MODULE
-	{
-		name = ODFC
-		MODE
-		{
-			MaxEC = 0.5
-			FUELS
-			{
-				MonoPropellant = 0.0005
-			}
-		}
-		MODE
-		{
-			MaxEC = 0.5
-			FUELS
-			{
-				LiquidFuel = 0.0005625
-				Oxidizer = 0.00152777777777778
-			}
-		}
-		MODE
-		{
-			MaxEC = 0.5
-			FUELS
-			{
-				LiquidFuel = 0.0000425
-				IntakeAir = 0.0102
-			}
-		}
-		MODE:NEEDS[CommunityResourcePack]
-		{
-			MaxEC = 0.5
-			FUELS
-			{
-				Hydrogen = 0.104264849617275
-				Oxygen = 0.0788621982550263
-			}
-			BYPRODUCTS
-			{
-				Water = 0.0000843419222
-			}
-		}
-	}
-
-	@RESOURCE[electricCharge]
-	{
-		&amount = 5
-		%amount += 5
-		&maxAmount = 5
-		%maxAmount += 5
-		flowState = False
-	}
-
-	@RESOURCE[MonoPropellant]
-	{
-		&amount = 5
-		%amount += 5
-		&maxAmount = 5
-		%maxAmount += 5
-	}
-}
-
-	// @RESOURCE[StoredCharge]:NEEDS[NearFutureElectrical]
-	// {
-		// &amount = 5
-		// %amount += 5
-		// &maxAmount = 5
-		// %maxAmount += 5
-	// }
-
-	// MODULE:NEEDS[NearFutureElectrical]
-	// {
-		// name = DischargeCapacitor
-		// ModuleID = lightColor
-		// ChargeRatio = .75 // 1
-		// ChargeRate = 0.1
-		// DischargeRate = 0.5
-		// MaximumCharge = 5
-	// }
-
-	// @RESOURCE[ReservePower]:NEEDS[AmpYear]
-	// {
-		// &amount = 5
-		// %amount += 5
-		// &maxAmount = 5
-		// %maxAmount += 5
-	// }
-
-	// @RESOURCE[megaJoules]:NEEDS[InterstellarFuelSwitch]
-	// {
-		// &amount = 0.01
-		// %amount += 0.01
-		// &maxAmount = 0.01
-		// %maxAmount += 0.01
-	// }
-
-// CC BY-NC-SA-4.0
-// zer0Kerbal
