--- GameData/OnDemandFuelCells/MM-Patches/Kethaane.txt.atom
@@ -1,92 +1,0 @@
-PART
-{
-<#569117>
-name = kethane_generator
-module = Part
-author = Keptin
-
-rescaleFactor = 1
-
-node_stack_top = 0.0, 0.40, 0.0, 0.0, 1.0, 0.0, 1
-node_stack_bottom = 0.0, -0.40, 0.0, 0.0, -1.0, 0.0, 1
-
-cost = 1900
-category = Utility
-subcategory = 0
-title = KE-G35 Kethoelectric Reactor Unit
-manufacturer = Organization of Kethane Equipment Producers
-description = Constructed from fragments of destroyed solar panels, the KE-G35 is a remarkably efficient Kethane-powered electrical generator. The exhaust manifold captures and stores some of the Xenon gas dissolved in the fuel, making this a versatile unit for exploration and drilling. The generator automatically regulates its output in order to save fuel.
-
-TechRequired = experimentalElectrics
-entryCost = 50
-
-attachRules = 1,0,1,0,0
-bulkheadProfiles = size1
-
-mass = 0.3
-dragModelType = default
-maximum_drag = 0.2
-minimum_drag = 0.3
-angularDrag = 2
-crashTolerance = 6
-breakingForce = 50
-breakingTorque = 50
-maxTemp = 2000
-fuelCrossFeed = True
-
-MODEL
-{
-   model = Kethane/Parts/kethane_generator/model
-   position = 0, -0.40, 0
-   scale = 0.61, 0.61, 0.61
-   rotation = 0, 0, 0
-}
-
-MODULE
-{
-    name = KethaneGenerator
-    KethaneRate = 0.5
-    PowerRate = 75
-    XenonMassRatio = 0.04
-    MaxEmission = 300
-    MinEmission = 100
-}
-
-MODULE:NEEDS[Kethane]
-{
-	MaxEC = 75
-	FUEL
-	{
-		Kethane	= 0.5
-	}
-	Byproducts
-	{
-		xenonGas = 0.04
-	}
-}
-}
-
-RESOURCE_DEFINITION
-{
-  name = Kethane
-  density = 0.002
-  unitCost = 0.16
-  flowMode = ALL_VESSEL
-  transfer = PUMP
-  volume = 2
-}
-
-RESOURCE_DEFINITION
-{
-  name = KIntakeAir
-  density = 0.001
-  flowMode = NO_FLOW
-  transfer = NONE
-}
-
-@TANK_DEFINITION[Default] {
-    TANK {
-        name = Kethane
-        fillable = false
-    }
-}
