--- GameData/OnDemandFuelCells/MM-Patches/Modify/DuoPods.cfg
@@ -1,103 +1,0 @@
-// DuoPods v1.
-// On Demand Fuel Cells Refueled
-// created: 19 Sep 19
-// updated: 19 Sep 19
-
-@PART[mk1a,mk1b,mk1c]:NEEDS[ODFC,KGEx/DuoPods]:AFTER[ODFC]
-{
-	// @name ^= :^:ODFC:
-	@cost += 50
-	@title ^= :$: <color=#6495ED> ODFC</color>:
-	@description ^= :(.)$:$0\n<#6495ED><i>ODFC installed.</i> </size></color>:
-	@mass += 0.001
-	@tags ^= :$:  odfc demand:
-
-	!MODULE[ODFC] {}
-	MODULE,0
-	{
-		name = ODFC
-		MODE
-		{
-			MaxEC = 5
-			FUELS
-			{
-				MonoPropellant = 0.005
-			}
-		}
-		MODE
-		{
-			MaxEC = 5
-			FUELS
-			{
-				LiquidFuel = 0.005625
-				Oxidizer = 0.0152777777777778
-			}
-		}
-		MODE
-		{
-			MaxEC = 5
-			FUELS
-			{
-				LiquidFuel = 0.000425
-				IntakeAir = 0.102
-			}
-		}
-		MODE:NEEDS[CommunityResourcePack]
-		{
-			MaxEC = 5
-			FUELS
-			{
-				Hydrogen = 1.04264849617275
-				Oxygen = 0.788621982550263
-			}
-			BYPRODUCTS
-			{
-				Water = 0.000843419222
-			}
-		}
-	}
-
-	@RESOURCE[electricCharge]
-	{
-		&amount = 5
-		%amount += 5
-		&maxAmount = 5
-		%maxAmount += 5
-		flowState = False
-	}
-
-	@RESOURCE[MonoPropellant]
-	{
-		&amount = 5
-		%amount += 5
-		&maxAmount = 5
-		%maxAmount += 5
-	}
-
-	@RESOURCE[StoredCharge]:NEEDS[NearFutureElectrical]
-	{
-		&amount = 5
-		%amount += 5
-		&maxAmount = 5
-		%maxAmount += 5
-	}
-
-	@RESOURCE[ReservePower]:NEEDS[AmpYear]
-	{
-		&amount = 5
-		%amount += 5
-		&maxAmount = 5
-		%maxAmount += 5
-	}
-
-	@RESOURCE[megaJoules]:NEEDS[InterstellarFuelSwitch]
-	{
-		&amount = 0.01
-		%amount += 0.01
-		&maxAmount = 0.01
-		%maxAmount += 0.01
-	}
-}
-
-// CC BY-NC-SA-4.0
-// zer0Kerbal
