--- GameData/OnDemandFuelCells/MM-Patches/Modify/JatwaaDemolitionsCo.cfg
@@ -1,78 +1,0 @@
-// JatwaaDemolitionsCo.cfg v1.1
-// On Demand Fuel Cells Refueled
-// created: 31 Aug 19
-// updated: 02 Sep 19
-
-@PART[electricityGenerator]:NEEDS[ODFC,KGEx/JatwaaDemolitionsCo]:AFTER[ODFC]
-{
-	// @name ^= :^:ODFC:
-	@cost += 50
-	@title ^= :$: <color=#6495ED> ODFC</color>:
-	@description ^= :$:<#6495ED><i>ODFC installed. </i></color>:
-	@mass += 0.001
-	@tags ^= :$:  odfc demand:
-
-	!MODULE[ModuleResourceConverter] {}
-	MODULE
-	{
-		name = ODFC
-		MODE
-		{
-			MaxEC = 18
-			FUELS
-			{
-				MonoPropellant = 0.018
-			}
-		}
-		MODE
-		{
-			MaxEC = 18
-			FUELS
-			{
-				LiquidFuel = 0.02025
-				Oxidizer = 0.055
-			}
-		}
-		MODE:NEEDS[CommunityResourcePack]
-		{
-			MaxEC = 18
-			FUELS
-			{
-				Hydrogen = 3.75353458622189
-				Oxygen = 2.83903913718094
-			}
-			BYPRODUCTS
-			{
-				Water = 0.00303630919748661
-			}
-		}
-		MODE
-		{
-			MaxEC = 18
-			FUELS
-			{
-				LiquidFuel = 0.00153
-				IntakeAir = 0.3672
-			}
-		}
-	}
-
-	%RESOURCE[electricCharge]
-	{
-		&amount = 5
-		%amount += 5
-		&maxAmount = 5
-		%maxAmount += 5
-		flowState = False
-	}
-
-	%RESOURCE[LiquidFuel]
-	{
-		&amount = 5
-		%amount += 5
-		&maxAmount = 5
-		%maxAmount += 5
-	}
-}
-// CC BY-NC-SA-4.0
-// zer0Kerbal
