--- GameData/OnDemandFuelCells/MM-Patches/Modify/SolidFuelCell.cfg
@@ -1,39 +1,0 @@
-// SolidFuelCell.cfg v1.1
-// On Demand Fuel Cells Refueled
-// created: 29 Aug 19
-// updated: 03 Sep 19
-
-@PART[solid_fuelcell]:NEEDS[ODFC,SolidFuelCell]:AFTER[ODFC]
-{
-	// @name ^= :^:ODFC:
-	@cost += 50
-	@title ^= :$: <color=#6495ED> ODFC</color>:
-	@description ^= :$:<#6495ED><i>ODFC installed. </i></color>:
-	@mass += 0.001
-	@tags ^= :$:  odfc demand:
-
-	!MODULE[ModuleResourceConverter] {}
-	MODULE
-	{
-		name = ODFC
-		MODE
-		{
-			MaxEC = 1.5
-			FUELS
-			{
-				SolidFuel = 0.00375
-			}
-		}
-		MODE
-		{
-			MaxEC = 1.5
-			FUELS
-			{
-				SolidFuel = 0.00375
-			}
-		}
-	}
-}
-
-// CC BY-NC-SA-4.0
-// zer0Kerbal
