--- GameData/OnDemandFuelCells/MM-Patches/Modify/UniversalStorage.cfg
@@ -1,62 +1,0 @@
-// UniversalStorage2.cfg v1.3
-// On Demand Fuel Cells Refueled
-// created: 22 Aug 19
-// updated: 01 Sep 19
-
-// NOTE: HOW mode seems way off (too little) based upon US2 stats. other modes use ODFC standard 1.0 numbers.
-
-@PART[US_1P110_Wedge_Fuelcell]:NEEDS[UniversalStorage,ODFC]:AFTER[ODFC]
-{
-	//@name ^= :^:ODFC:
-	@title ^= :$: <color=green> ODFC</color>:
-	@description ^= :$: <color=green> ODFC version. </color>:
-	@tags ^= :$:  odfc demand:
-
-	!MODULE[ModuleResourceConverter] {}
-	MODULE {
-		name = ODFC
-		MODE
-		{
-			MaxEC = 16
-			FUELS
-			{
-				MonoPropellant = 0.016
-			}
-		}
-		MODE:NEEDS[CommunityResourcePack]
-		{
-			MaxEC = 16
-			FUELS
-			{
-				Hydrogen = 0.064736	// Hydrogen = 2.7298433354341
-				Oxygen = 0.032368		// Oxidizer = 0.0488888888888889
-			}
-			BYPRODUCTS
-			{
-				Water = 0.0000520721	// Water = 0.00220822487089936
-			}
-		}
-		MODE
-		{
-			MaxEC = 16
-			FUELS
-			{
-				LiquidFuel = 0.018
-				Oxidizer = 0.0488888888888889
-
-			}
-		}
-		MODE
-		{
-			MaxEC = 16
-			FUELS
-			{
-				LiquidFuel = 0.00136
-				IntakeAir = 0.3264
-			}
-		}
-	}
-}
-
-// CC BY-NC-SA-4.0
-// zer0Kerbal
