--- GameData/OnDemandFuelCells/MM-Patches/Modify/UniversalStorage2.cfg
@@ -1,116 +1,0 @@
-// UniversalStorage2.cfg v1.4
-// On Demand Fuel Cells Refueled
-// created: 22 Aug 19
-// updated: 03 Sep 19
-
-// large fuel cell gains efficiency ie uses less fuel (just in HO mode) as compared to the small fuel cell.
-@PART[USFuelCellSmal]:NEEDS[UniversalStorage2,ODFC]:AFTER[ODFC]
-{
-	// @name ^= :^:ODFC:
-	@cost += 50
-	@title ^= :$: <color=#6495ED> ODFC</color>:
-	@description ^= :$:<#6495ED><i>ODFC installed. </i></color>:
-	@mass += 0.001
-	@tags ^= :$:  odfc demand:
-
-	!MODULE[ModuleResourceConverter] {}
-	MODULE {
-		name = ODFC
-		MODE
-		{
-			MaxEC = 1.6
-			FUELS
-			{
-				MonoPropellant = 0.0016
-			}
-		}
-		MODE:NEEDS[CommunityResourcePack]
-		{
-			MaxEC = 1.6
-			FUELS
-			{
-				Hydrogen = 0.33364751877527900000
-				Oxygen = 0.16788377319816400000
-			}
-			BYPRODUCTS
-			{
-				Water = 0.00026989415088769900
-			}
-		}
-		MODE
-		{
-			MaxEC = 1.6
-			FUELS
-			{
-				LiquidFuel = 0.0018
-				Oxidizer = 0.004888889
-			}
-		}
-		MODE
-		{
-			MaxEC = 1.6
-			FUELS
-			{
-				LiquidFuel = 0.000136
-				IntakeAir = 0.03264
-			}
-		}
-	}
-}
-
-@PART[USFuelCellMedium]:NEEDS[UniversalStorage2,ODFC]:AFTER[ODFC]
-{
-	// @name ^= :^:ODFC:
-	@cost += 50
-	@title ^= :$: <color=#6495ED> ODFC</color>:
-	@description ^= :$:<#6495ED><i>ODFC installed. </i></color>:
-	@mass += 0.001
-	@tags ^= :$:  odfc demand:
-
-	!MODULE[ModuleResourceConverter] {}
-	MODULE {
-		name = ODFC
-		MODE:NEEDS[CommunityResourcePack]
-		{
-			MaxEC = 5
-			FUELS
-			{
-				Hydrogen = 0.42924828374530600000
-				Oxygen = 0.85307604232315700000
-			}
-			BYPRODUCTS
-			{
-				Water = 0.00069007027215604800
-			}
-		}
-		MODE
-		{
-			MaxEC = 5
-			FUELS
-			{
-				MonoPropellant = 0.005
-			}
-		}
-		MODE
-		{
-			MaxEC = 5
-			FUELS
-			{
-				LiquidFuel = 0.005625
-				Oxidizer = 0.015277778
-			}
-		}
-		MODE
-		{
-			MaxEC = 5
-			FUELS
-			{
-				LiquidFuel = 0.000425
-				IntakeAir = 0.102
-			}
-		}
-	}
-}
-
-// CC BY-NC-SA-4.0
-// zer0Kerbal
