--- GameData/OnDemandFuelCells/MM-Patches/SETIprobeParts-ProbeCores.cfg
@@ -1,118 +1,0 @@
-// SETIprobeParts-ProbeCores v1.0
-// On Demand Fuel Cells Refueled
-// created: 13 Sep 19
-// updated: 13 Sep 19
-
-@PART[setiProbeStack1,setiProbeStack2]:NEEDS[ODFC,SETIprobeParts]:AFTER[ODFC]
-{
-	// @name ^= :^:ODFC:
-	@cost += 5
-	@title ^= :$: <#6495ED> ODFC</color>:
-	@description ^= :$:<#6495ED><i>ODFC installed. </i></color>:
-	@mass += 0.001
-	tags = seti probe alkaline fuel cell mini odfc demand command pod core
-
-	!MODULE[ModuleResourceConverter] {}
-	MODULE,0
-	{
-		name = ODFC
-		MODE:NEEDS[CommunityResourcePack]
-		{
-			MaxEC = .2
-			FUELS
-			{
-				Hydrogen = 0.0417059398469099
-				Oxygen = 0.0315448793020105
-			}
-			BYPRODUCTS
-			{
-				Water = 0.0000337367688609624
-			}
-		}
-		MODE
-		{
-			MaxEC = .2
-			FUELS
-			{
-				MonoPropellant = 0.0002
-			}
-		}
-		MODE
-		{
-			MaxEC = .2
-			FUELS
-			{
-				LiquidFuel = 0.000225
-				Oxidizer = 0.000611111111111111
-			}
-		}
-		MODE
-		{
-			MaxEC = .2
-			FUELS
-			{
-				LiquidFuel = 0.000017
-				IntakeAir = 0.00408
-			}
-		}
-	}
-
-	@RESOURCE[Hydrogen]:NEEDS[CommunityResourcePack]
-	{
-	// Units: Liters
-	// Stored as a cryogenic liquid (expansion ratio 851)
-		amount = 15674
-		maxAmount = 15674
-	}
-
-	@RESOURCE[Oxygen]:NEEDS[CommunityResourcePack]
-	{
-	// Units: Liters
-	// Stored as a cryogenic liquid (expansion ratio 860)
-		amount = 7995
-		maxAmount = 7995
-	}
-
-	@RESOURCE[electricCharge]
-	{
-		&amount = 5
-		%amount += 5
-		&maxAmount = 5
-		%maxAmount += 5
-		flowState = False
-	}
-
-	@RESOURCE[MonoPropellant]
-	{
-		&amount = 5
-		%amount += 5
-		&maxAmount = 5
-		%maxAmount += 5
-	}
-
-	@RESOURCE[StoredCharge]:NEEDS[NearFutureElectrical]
-	{
-		&amount = 5
-		%amount += 5
-		&maxAmount = 5
-		%maxAmount += 5
-	}
-
-	// @RESOURCE[ReservePower]:NEEDS[AmpYear]
-	// {
-		// &amount = 5
-		// %amount += 5
-		// &maxAmount = 5
-		// %maxAmount += 5
-	// }
-
-	// @RESOURCE[megaJoules]:NEEDS[InterstellarFuelSwitch]
-	// {
-		// &amount = 0.01
-		// %amount += 0.01
-		// &maxAmount = 0.01
-		// %maxAmount += 0.01
-	// }
-}
-// CC BY-NC-SA-4.0
-// zer0Kerbal
