--- GameData/OnDemandFuelCells/MM-Patches/SETIprobeParts.cfg
@@ -1,76 +1,0 @@
-// SETIprobeParts v1.0
-// On Demand Fuel Cells Refueled
-// created: 13 Sep 19
-// updated: 13 Sep 19
-
-@PART[setiMiniAFC]:NEEDS[ODFC,SETIprobeParts]:AFTER[ODFC]
-{
-	// @name ^= :^:ODFC:
-	@cost += 5
-	@title ^= :$: <#6495ED> ODFC</color>:
-	@description ^= :$:<#6495ED><i>ODFC installed. </i></color>:
-	@mass += 0.001
-	tags = seti probe alkaline fuel cell mini odfc demand
-
-	!MODULE[ModuleResourceConverter] {}
-	MODULE,0
-	{
-		name = ODFC
-		MODE
-		{
-			MaxEC = 2
-			FUELS
-			{
-				MonoPropellant = 0.002
-			}
-		}
-		MODE
-		{
-			MaxEC = 2
-			FUELS
-			{
-				LiquidFuel = 0.00225
-				Oxidizer = 0.00611111111111111
-			}
-		}
-		MODE:NEEDS[CommunityResourcePack]
-		{
-			MaxEC = 2
-			FUELS
-			{
-				Hydrogen = 0.417059398469099
-				Oxygen = 0.315448793020105
-			}
-			BYPRODUCTS
-			{
-				Water = 0.000337367688609624
-			}
-		}
-		MODE
-		{
-			MaxEC = 2
-			FUELS
-			{
-				LiquidFuel = 0.00017
-				IntakeAir = 0.0408
-			}
-		}
-		MODE:NEEDS[CommunityResourcePack,RealFuels]
-		{
-			MaxEC = 2
-			FUELS
-			{
-				Hydrazine = 0.03128
-				Oxygen = 0.02364
-			}
-			BYPRODUCTS
-			{
-				Water = 0.00002532
-			}
-		}
-	}
-}
-
-
-// CC BY-NC-SA-4.0
-// zer0Kerbal
