--- GameData/OnDemandFuelCells/ODFCCopyPatches.version
@@ -1,37 +1,0 @@
-{
-	"NAME":"On Demand Fuel Cells (ODFC) - Copy Patches",
-	"URL":"https://raw.githubusercontent.com/zer0Kerbal/ODFCr/master/GameData/ODFCCopyPatches.version",
-	"DOWNLOAD":"https://github.com/zer0Kerbal/ODFCr/releases", 
-	"CHANGE_LOG_URL": "https://raw.githubusercontent.com/zer0Kerbal/ODFCr/master/GameData/CopyPatchesChangelog.md",
-    "GITHUB":
-    {
-        "USERNAME":"zer0Kerbal",
-        "REPOSITORY":"OnDemandFuelCells",
-        "ALLOW_PRE_RELEASE":false,
-    },
-	"VERSION":
-	{
-		"MAJOR":1,
-		"MINOR":0,
-		"PATCH":0,
-		"BUILD":0
-	},
-	"KSP_VERSION":
-	{
-		"MAJOR":1,
-		"MINOR":7,
-		"PATCH":3
-	},
-	"KSP_VERSION_MIN":
-	{
-		"MAJOR":1,
-		"MINOR":4,
-		"PATCH":1
-	},
-	"KSP_VERSION_MAX":
-	{
-		"MAJOR":1,
-		"MINOR":8,
-		"PATCH":9999
-	}
-}
