--- GameData/OnDemandFuelCells/ODFCModifyPatchesReadme.txt
@@ -1,8 +1,0 @@
-﻿# Modify Patches for On Demand Fuel Cells
-
-One or both sets of patches needed for ODFC
-
-see ODFC changelog for changes.
-
-CC BY-NC-SA-4.0
-zer0Kerbal
