+++ GameData/OnDemandFuelCells/OnDemandFuelCells.version
@@ -6,14 +6,14 @@
     {
         "USERNAME":"zer0Kerbal",
         "REPOSITORY":"OnDemandFuelCells",
-        "ALLOW_PRE_RELEASE":false,
+        "ALLOW_PRE_RELEASE":false
     },
 	"VERSION":
 	{
 		"MAJOR":1,
 		"MINOR":1,
 		"PATCH":2,
-		"BUILD":0
+		"BUILD":1
 	},
 	"KSP_VERSION":
 	{
