--- GameData/OnDemandFuelCells/PatchesChangelog.md
@@ -1,20 +1,0 @@
-# ODFC modify part patches  
-## for On Demand Fuel Cells Refueled (ODFCr)  
-
-### Changelog
-#### STATUS:
- * ***BETA PRE-Release***
-
-	VERSION
-	{
-		version = v.1.0.0.0
-		change = Initial release (for KSP v1.7.x)
-		change = added (@PART) patches for: 
-		change = Stock Mining Expansion
-		change = Stock
-		change = Stock Pods
-		change = Universal Storage
-		change = Universal Storage II
-		change = JatwaaDemolitionsCo
-	}
-}
