+++ GameData/OnDemandFuelCells/changelog.md
@@ -4,6 +4,12 @@
 ### Changelog
 #### STATUS:
  * ***Release***
+#### 1.1.2.1 - insert title HotBeverageReheated
+ * [BUG 1.1.2.0a] TweakScale
+ * [NEW][BUG 1.1.2.1a] AmpYear doesn't seem to recognize ODFC
+ * [NEW][BUG 1.1.2.0b] Kerbalism is not compatible with ODFC - Kerbalism developers have chosen to not integrate since they have their own version.
+ * [D][BUG 1.1.2.1b] Found the problem, Bill left a torque converter, which was set to moar, in the innards. Plus Val thought it would be fun to hide her signature in the code of the FCOS (Fuel Cell Operating System). In doing so, she overwrote the code to remember the editor settings when translating over to flight. This has been fixed by moving her signature into another mod's code. She's still happy.
+
 #### 1.1.2.0 - Stall Alert! aka Stall Tactics
  * ***[NEW]*** Optional difficulty settings for moderate and hard play - stall. Don't run out of EC or the fuel cells will stop functioning since they need electricity to function. Thank you to LinuxGuruGamer for the Setting.cs code and assistance.
  * ***[NEW]*** feature: autoSwitch (disabled for hard play). If current fuel mode is lacking fuel, will automatically look for another that has fuel.
@@ -13,7 +19,7 @@
  * continued working on [BUG 0.0.1.9b]
  * /*CODE*/ udfs - updateFS
  * /*CODE*/ udft - updateFT
- * [NEW][BUG 1.1.2.0a] Tweakscale will not scale module ODFC because onLoad() vs onStart issues?
+ * [NEW][BUG 1.1.2.0a] TweakScale will not scale module ODFC because onLoad() vs onStart issues?
  * reversed removal of color coding PAW label. those coders have been sacked, and those sackers of the previously sacked are now extra-judicially *sacked*
  * PAW label now acts as a color coded STATUS label - dynamically updating
  * [NEW][BUG 1.1.2.0b] Kerbalism doesn't seem to recognize the unique brilliance of ODFC, not their fault hard to see when its that bright. :D
