+++ OnDemandFuelCells.version
@@ -6,7 +6,7 @@
     {
         "USERNAME":"zer0Kerbal",
         "REPOSITORY":"OnDemandFuelCells",
-        "ALLOW_PRE_RELEASE":false,
+        "ALLOW_PRE_RELEASE":false
     },
 	"VERSION":
 	{
