--- OnDemandFuelCells/AssemblyVersion.cs
@@ -1,1 +1,0 @@
-﻿ErrorGeneratingOutput
