+++ Source/AssemblyVersion.tt
@@ -2,7 +2,6 @@
 <#@ import namespace="System.IO" #>
 <#@ output extension=".cs" #>
 
-<#@ assembly name="EnvDTE" #><# /* This assembly provides access to Visual Studio project properties. */ #>
 <#
 
 	// Instructions
@@ -10,7 +9,7 @@
 	// 2.  Copy this file into the new template
 	// 3.  Update the string: versionfile with the complete path to the .version file
 	// 4.  Remove the following line from the file AssemblyInfo.cs (usually located in the "Property" folder inside your C# project):
-	//        [assembly: AssemblyFileVersion("1.0.0.0")]
+	//        [assembly: AssemblyVersion("1.0.0.0")]
 	// 5.  Add the following to the PreBuild steps:
 	//		set textTemplatingPath="%CommonProgramFiles(x86)%\Microsoft Shared\TextTemplating\$(VisualStudioVersion)\texttransform.exe"
 	//
@@ -18,26 +17,24 @@
 	//
 	//		%textTemplatingPath% "$(ProjectDir)AssemblyFileVersion.tt"
 
-    int major = 0; 
-    int minor = 0; 
-    int build = 0; 
+    int major = 0;
+    int minor = 0;
+    int build = 0;
 	int patch = 0;
 	bool versionSection = false;
+
 	int i = 0;
 	int i2 = 0;
 	string s;
 
-		// For Visual Studio / MSBuild Build-Time Template Resolution
-	string RootDirectory = System.IO.Path.GetDirectoryName(Host.TemplateFile) + @"\..\";
-
 	//
-	// Update the following with the name of the .version file which is in the root directory
+	// Update the following with the complete path to the .version file
 	//
-	string versionfile = RootDirectory + "OnDemandFuelCells.version";
+	string versionfile = @"C:\KSP_DEV\Workspace\ODFCr\OnDemandFuelCells.version";
 
 	if (!File.Exists(versionfile))
 	{
-		Write("File: " + versionfile + " missing\n");		
+		Write("File: " + versionfile + " missing\n");
 	}
 
 	try
@@ -63,37 +60,37 @@
 					{
 						s = line.Substring(i + 1, i2 - i - 1);
 
-						if (line.Contains("MAJOR"))			
-							Int32.TryParse(s, out major);			
-			
-						if (line.Contains("MINOR"))			
+						if (line.Contains("MAJOR"))
+							Int32.TryParse(s, out major);
+
+						if (line.Contains("MINOR"))
 							Int32.TryParse(s, out minor);
-			
-						if (line.Contains("PATCH"))			
+
+						if (line.Contains("PATCH"))
 							Int32.TryParse(s, out patch);
-			
-						if (line.Contains("BUILD"))			
+
+						if (line.Contains("BUILD"))
 							Int32.TryParse(s, out build);
 					}
 				}
 			}
 		}
-         
+
      }
      catch
      {
-         major = 1; 
-         minor = 0; 
-         patch = 0; 
-         build = 0; 
+         major = 1;
+         minor = 0;
+         patch = 0;
+         build = 0;
      }
 	 //Write("File done");
 
  #>
  // This code was generated by a tool. Any changes made manually will be lost
  // the next time this code is regenerated.
- // 
-  
+ //
+
  using System.Reflection;
 
  [assembly: AssemblyVersion("<#= major #>.<#= minor #>.<#= patch #>.<#= build #>")]
