+++ Source/InstallChecker.cs
