+++ Source/OnDemandFuelCells.csproj
@@ -1,5 +1,5 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
-<Project ToolsVersion="15.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+<Project ToolsVersion="15.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
   <PropertyGroup>
     <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
@@ -11,8 +11,6 @@
     <AssemblyName>ODFC</AssemblyName>
     <TargetFrameworkVersion>v3.5</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
-    <ReleaseVersion>1.1.1.0</ReleaseVersion>
-    <Deterministic>true</Deterministic>
   </PropertyGroup>
   <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
     <DebugSymbols>true</DebugSymbols>
@@ -23,34 +21,32 @@
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
   </PropertyGroup>
-  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
+  <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Release|AnyCPU'">
     <DebugType>pdbonly</DebugType>
     <Optimize>true</Optimize>
     <OutputPath>bin\Release\</OutputPath>
     <DefineConstants>TRACE</DefineConstants>
     <ErrorReport>prompt</ErrorReport>
     <WarningLevel>4</WarningLevel>
+    <DebugSymbols>true</DebugSymbols>
   </PropertyGroup>
-  <!-- <PropertyGroup> -->
-  <!-- <RootNamespace>ODFC</RootNamespace> -->
-  <!-- </PropertyGroup -->
   <ItemGroup>
-    <Reference Include="System" />
-    <Reference Include="System.Core" />
-    <Reference Include="System.Data" />
-    <Reference Include="System.Data.DataSetExtensions" />
-    <Reference Include="System.Xml.Linq" />
-    <Reference Include="System.Xml" />
     <Reference Include="Assembly-CSharp">
       <HintPath>$(DevDir)\KSP_x64_Data\Managed\Assembly-CSharp.dll</HintPath>
       <Private>False</Private>
     </Reference>
-    <Reference Include="Assembly-CSharp-firstpass">
-      <HintPath>$(DevDir)\KSP_x64_Data\Managed\Assembly-CSharp-firstpass.dll</HintPath>
-      <Private>False</Private>
-    </Reference>
+    <Reference Include="System" />
+    <Reference Include="System.Core" />
+    <Reference Include="System.Xml.Linq" />
+    <Reference Include="System.Data.DataSetExtensions" />
+    <Reference Include="System.Data" />
+    <Reference Include="System.Xml" />
     <Reference Include="UnityEngine">
       <HintPath>$(DevDir)\KSP_x64_Data\Managed\UnityEngine.dll</HintPath>
+      <Private>False</Private>
+    </Reference>
+    <Reference Include="Assembly-CSharp-firstpass">
+      <HintPath>$(DevDir)\KSP_x64_Data\Managed\Assembly-CSharp-firstpass.dll</HintPath>
       <Private>False</Private>
     </Reference>
     <Reference Include="UnityEngine.UI">
@@ -59,34 +55,34 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="InstallChecker.cs" />
+    <Compile Include="AssemblyVersion.cs">
+      <AutoGen>True</AutoGen>
+      <DesignTime>True</DesignTime>
+      <DependentUpon>AssemblyVersion.tt</DependentUpon>
+    </Compile>
     <Compile Include="ODFC.cs" />
     <Compile Include="Config.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="InstallChecker.cs" />
     <Compile Include="Settings.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <Content Include="AssemblyVersion.tt">
+      <Generator>TextTemplatingFileGenerator</Generator>
+      <LastGenOutput>AssemblyVersion.cs</LastGenOutput>
+    </Content>
+  </ItemGroup>
+  <ItemGroup>
+    <Service Include="{508349B6-6B84-4DF5-91F0-309BEEBAD82D}" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
-    <PostBuildEvent>
-      "C:\KSP_DEV\Workspace\pdb2mdb.exe" "C:\KSP_DEV\Workspace\ODFCr\OnDemandFuelCells\bin\Release\ODFC.dll"
-      xcopy /Y "C:\KSP_DEV\Workspace\ODFCr\OnDemandFuelCells\bin\Release\ODFC.dll" "C:\KSP_DEV\Workspace\ODFCr\GameData\OnDemandFuelCells\Plugins"
-      xcopy /Y "C:\KSP_DEV\Workspace\ODFCr\OnDemandFuelCells\bin\Release\ODFC.dll.mdb" "C:\KSP_DEV\Workspace\ODFCr\GameData\OnDemandFuelCells\Plugins"
-      xcopy /Y "C:\KSP_DEV\Workspace\ODFCr\CC BY-NC-SA-4.0.txt" "C:\KSP_DEV\Workspace\ODFCr\GameData\OnDemandFuelCells"
-      xcopy /Y "C:\KSP_DEV\Workspace\ODFCr\changelog.md" "C:\KSP_DEV\Workspace\ODFCr\GameData\OnDemandFuelCells"
-      xcopy /Y "C:\KSP_DEV\Workspace\ODFCr\changelog_plugin.txt" "C:\KSP_DEV\Workspace\ODFCr\GameData\OnDemandFuelCells"
-      xcopy /Y "C:\KSP_DEV\Workspace\ODFCr\license.md" "C:\KSP_DEV\Workspace\ODFCr\GameData\OnDemandFuelCells"
-      xcopy /Y "C:\KSP_DEV\Workspace\ODFCr\OnDemandFuelCells.version" "C:\KSP_DEV\Workspace\ODFCr\GameData\OnDemandFuelCells"
-      xcopy /Y "C:\KSP_DEV\Workspace\ODFCr\README.md" "C:\KSP_DEV\Workspace\ODFCr\GameData\OnDemandFuelCells"
-    </PostBuildEvent>
+    <PostBuildEvent>start /D $(ProjectDir)..\  /WAIT deploy.bat  $(TargetDir) $(TargetFileName)
+if $(ConfigurationName) == Release start /D $(ProjectDir)..\   /WAIT buildRelease.bat $(TargetDir) $(TargetFileName)</PostBuildEvent>
   </PropertyGroup>
   <PropertyGroup>
-    <PreBuildEvent>
-      <!-- set textTemplatingPath=\texttransform.exe" -->
-      <!-- $(textTemplatingPath)\TextTransform.exe "$(ProjectDir)AssemblyVersion.tt" -->
-      <!--if %textTemplatingPath%=="\Microsoft Shared\TextTemplating\$(VisualStudioVersion)\texttransform.exe" set textTemplatingPath="%CommonProgramFiles%\Microsoft Shared\TextTemplating\$(VisualStudioVersion)\texttransform.exe"
-      set textTemplatingPath="%CommonProgramFiles(x86)%\Microsoft Shared\TextTemplating\$(VisualStudioVersion)\texttransform.exe"
-      %textTemplatingPath% "$(ProjectDir)AssemblyFileVersion.tt"-->
-    </PreBuildEvent>
+    <PreBuildEvent>set textTemplatingPath="$(DevEnvDir)texttransform.exe"
+%25textTemplatingPath%25 "$(ProjectDir)AssemblyVersion.tt"</PreBuildEvent>
   </PropertyGroup>
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it. 
        Other similar extension points exist, see Microsoft.Common.targets.
