+++ Source/Settings.cs
