+++ changelog.md
@@ -4,8 +4,12 @@
 ### Changelog
 #### STATUS:
  * ***Release***
-#### 1.1.2.1 - insert title HotBeverageReheated
+#### 1.1.2.1 - Set SCE to AUX
  * [BUG 1.1.2.0a] TweakScale
+ * [NEW][BUG 1.1.2.1a] AmpYear doesn't seem to recognize ODFC
+ * BUG 1.1.2.0b] Kerbalism is not compatible with ODFC - Kerbalism developers have chosen to not integrate since they have their own version.
+ * [D][BUG 1.1.2.1b] Found the problem, Bill left a torque converter, which was set to moar, in the innards. Plus Val thought it would be fun to hide her signature in the code of the FCOS (Fuel Cell Operating System). In doing so, she overwrote the code to remember the editor settings when translating over to flight. This has been fixed by moving her signature into another mod's code. She's still happy.
+ * Updated to automated build (deploy.bat,buildRelease.bat/Assembly.tt)
 
 #### 1.1.2.0 - Stall Alert! aka Stall Tactics
  * ***[NEW]*** Optional difficulty settings for moderate and hard play - stall. Don't run out of EC or the fuel cells will stop functioning since they need electricity to function. Thank you to LinuxGuruGamer for the Setting.cs code and assistance.
@@ -24,7 +28,7 @@
 ####  0.0.1.9 (this is actually 1.1.1.9 and next release will switch to 1.1.2.0)
 0.0.1.9 (this is actually 1.1.1.9 and next release will switch to 1.1.2.0)
  * added item grouping in PAW.
- * [NEW][BUG 0.0.1.9a] - B9 module swapping - needs onLoad etc update to make work
+ * [NEW][BUG 0.0.1.9a] B9 module swapping - needs onLoad etc update to make work
  * [NEW][BUG 0.0.1.9b] next fuel mode should not be visible when only one mode
  * [D][BUG 0.0.1.9c] mangled config caused this. added error checking in cfg.cs - thank you LGG for this code.
  * [NEW][BUG 0.0.1.9c] ERROR!'s out when there is only one fuel mode. Stock pod patch only adds one mode (monoprop - because pods usually have monoprop if they have any fuel). This bug was temporarily fixed by added a second mode(it can be the same as the first so it appears like there only one fuel mode) in the patch (LFO).
