+++ docs/postbuildevents.txt.atom
@@ -1,0 +1,10 @@
+
+"C:\KSP_DEV\Workspace\pdb2mdb.exe" "C:\KSP_DEV\Workspace\ODFCr\OnDemandFuelCells\bin\Release\ODFC.dll"
+xcopy /Y "C:\KSP_DEV\Workspace\ODFCr\OnDemandFuelCells\bin\Release\ODFC.dll" "C:\KSP_DEV\Workspace\ODFCr\GameData\OnDemandFuelCells\Plugins"
+xcopy /Y "C:\KSP_DEV\Workspace\ODFCr\OnDemandFuelCells\bin\Release\ODFC.dll.mdb" "C:\KSP_DEV\Workspace\ODFCr\GameData\OnDemandFuelCells\Plugins"
+xcopy /Y "C:\KSP_DEV\Workspace\ODFCr\CC BY-NC-SA-4.0.txt" "C:\KSP_DEV\Workspace\ODFCr\GameData\OnDemandFuelCells"
+xcopy /Y "C:\KSP_DEV\Workspace\ODFCr\changelog.md" "C:\KSP_DEV\Workspace\ODFCr\GameData\OnDemandFuelCells"
+xcopy /Y "C:\KSP_DEV\Workspace\ODFCr\changelog_plugin.txt" "C:\KSP_DEV\Workspace\ODFCr\GameData\OnDemandFuelCells"
+xcopy /Y "C:\KSP_DEV\Workspace\ODFCr\license.md" "C:\KSP_DEV\Workspace\ODFCr\GameData\OnDemandFuelCells"
+xcopy /Y "C:\KSP_DEV\Workspace\ODFCr\OnDemandFuelCells.version" "C:\KSP_DEV\Workspace\ODFCr\GameData\OnDemandFuelCells"
+xcopy /Y "C:\KSP_DEV\Workspace\ODFCr\README.md" "C:\KSP_DEV\Workspace\ODFCr\GameData\OnDemandFuelCells"
