+++ docs/spacedockVersion-v1121.htm
@@ -1,0 +1,42 @@
+<!-- SpaceDock-version.htm v1.1.2.1
+On Demand Fuel Cells Refueled
+created: 17 Aug 19
+updated: 21 Sep 19 -->
+<h1>On Demand Fuel Cells Refueled</h1>
+![OndDemandFuelCellsRefueled 1.1.2.1](https://img.shields.io/badge/MOD%20version-1.1.2.1-orange.svg?style=flat-square) ![OndDemandFuelCellsRefueled](https://img.shields.io/badge/KSP%20version-1.7.x-66ccff.svg?style=flat-square)  ![OndDemandFuelCellsRefueled](https://img.shields.io/badge/CKAN-Indexed-brightgreen.svg)![Mod Version(manual)](https://img.shields.io/badge/MOD%20version-0.0.0.1-orange.svg?style=flat-square) ![CC 4.0 BY-NC-SA](https://img.shields.io/badge/license-CC--4.0--BY--SA-lightgrey)
+
+<p></p>
+<h3> v.1.1.2.1 -Set SCE to AUX</h3>
+<p>*See ![ChangeLog](https://github.com/zer0Kerbal/ODFCr/blob/master/changelog.md) for full details of mod changes*
+ #### 1.1.2.1 - Set SCE to AUX
+STATUS: *Release*
+  * [NEW]
+  * [NEW]
+  * [D]
+  <ul { list-style-type: square;}>
+    <li><b>[NEW]</b>[BUG 1.1.2.1a] AmpYear doesn't seem to recognize ODFC.</li>
+    <li><b>[NEW]</b>[BUG 1.1.2.0b] Kerbalism is not compatible with ODFC - Kerbalism developers have chosen to not integrate since they have their own version.</li>
+    <li><b>[NEW]</b>the group label on the PAW/RMB part window now reflects status of the fuel cell so you don't have to expand the group to peek in to see if the refrigerator light is on. thank you to @Blowfish, @Dmagic, and @LinuxGuruGamer for the assist. That is one LONG piece of code.</li>
+    <li><b>[D][NEW]</b>[BUG 1.1.2.1b] Found the problem, Bill left a torque converter, which was set to moar, in the innards. Plus Val thought it would be fun to hide her signature in the code of the FCOS (Fuel Cell Operating System). In doing so, she overwrote the code to remember the editor settings when translating over to flight. This has been fixed by moving her signature into another mod's code. She's still happy.</li>
+    <li>continued working on [BUG 1.1.2.0a]</li>
+    <li>Known Issue Tracker</li>
+    <ul { list-style-type: square;}>
+      <li>[NEW][BUG 1.1.2.1a] AmpYear doesn't seem to recognize ODFC</li>
+      <li>[BUG 1.1.2.0a] TweakScale will not scale module ODFC</li>
+      <li>[BUG 0.0.1.9a] B9 module swapping - needs onLoad etc update to make work</li>
+      <li>[BUG 0.0.1.9b] next fuel mode should not be visible when only one mode</li>
+      <li>[BUG 0.0.1.6a] Does not seeming work with BackgroundProcessing or Background Resources mods (being looked at) (so ODFC doesn't work when doesn't have focus). Should not have both BackgroundProcessing and BackgroundResources installed.</li>
+      <li>[BUG 1.1.2.0b] Kerbalism doesn't recognize ODFC, ODFC still functions correctly</li>
+      <li></li>
+    </ul>
+  </ul>
+</p>
+<hr>
+<p style="color: #FF0000;">Kerbal Space Program 1.7.3</p>
+<p style="color: #FF0000;">Unity 2017.1.3p1</p>
+<p style="color: #FF0000;">.NET 3.5</p>
+
+<!--
+CC BY-NC-SA-4.0
+zer0Kerbal
+-->
